// backend/src/ai_history/dto/ai_history.response.dto.ts
import { ApiProperty } from '@nestjs/swagger';
import { IsNotEmpty, IsString } from 'class-validator';

export class AiHistoryResponseDto {
    @ApiProperty({
        description: "The message generated by the AI (plain text or a question asked to the user)",
        type: 'string',
        example: "What step would you like to add to your roadmap?"
    })
    @IsString()
    @IsNotEmpty()
    message: string;

    @ApiProperty({
        description: "The user's or chatbot's response (always a string)",
        type: 'string',
        example: "I want to add a step for onboarding."
    })
    @IsString()
    response?: string;
}